# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API GameDashboard",
    "version": "1.0.0",
    "description": "Panel de gestion de serveurs de jeu. Les erreurs suivent Problem Details (RFC 9457) : un corps JSON portant `title`, `status` et `detail`, et non un simple texte."
  },
  "servers": [
    {
      "url": "https://panel.example",
      "description": "Le panel, sur son domaine"
    }
  ],
...
  "paths": {
...
    "/api/v1/client/servers/{server}/snapshots": {
      "get": {
        "summary": "Lister les instantanés du serveur sur son node (nom, date, cause, expiration, épingle) et l'état de la fonction (meta : système de fichiers, espace libre, écritures acceptées, épingles, prochaine demande manuelle possible). 404 si aucun agent du node n'offre les instantanés.",
        "description": "Clé personnelle. Portée requise : `snapshots.read`.",
        "tags": [
          "Instantanés"
        ],
        "operationId": "client-get-servers-by-server-snapshots",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": [
              "snapshots.read"
            ]
          }
        ],
        "x-gd-scope": "snapshots.read",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « snapshots.read », ou refusé pour la ressource visée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "post": {
        "summary": "Demander un instantané manuel. Répond 202 avec l'ordre : l'agent du node le prend à son relevé suivant, et une demande déjà en attente est rendue telle quelle. Refusé (409) si l'agent ne répond plus, si la fonction ou les demandes manuelles sont coupées sur ce node ou si l'espace libre manque, (429) avant la fin du délai entre deux demandes.",
        "description": "Clé personnelle. Portée requise : `snapshots.create`.",
        "tags": [
          "Instantanés"
        ],
        "operationId": "client-post-servers-by-server-snapshots",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": [
              "snapshots.create"
            ]
          }
        ],
        "x-gd-scope": "snapshots.create",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « snapshots.create », ou refusé pour la ressource visée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      }
    },
    "/api/v1/client/servers/{server}/snapshots/{snapshot}/pin": {
      "post": {
        "summary": "Épingler un instantané : { label? }. La rotation ne le détruit plus, jusqu'à la durée maximale du node. Refusé (409) au-delà de la limite d'épingles du serveur.",
        "description": "Clé personnelle. Portée requise : `snapshots.create`.",
        "tags": [
          "Instantanés"
        ],
        "operationId": "client-post-servers-by-server-snapshots-by-snapshot-pin",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "required": true,
            "description": "Valeur de « snapshot ».",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": [
              "snapshots.create"
            ]
          }
        ],
        "x-gd-scope": "snapshots.create",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « snapshots.create », ou refusé pour la ressource visée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      },
      "delete": {
        "summary": "Désépingler un instantané : la rotation le reprend.",
        "description": "Clé personnelle. Portée requise : `snapshots.create`.",
        "tags": [
          "Instantanés"
        ],
        "operationId": "client-delete-servers-by-server-snapshots-by-snapshot-pin",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "required": true,
            "description": "Valeur de « snapshot ».",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": [
              "snapshots.create"
            ]
          }
        ],
        "x-gd-scope": "snapshots.create",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « snapshots.create », ou refusé pour la ressource visée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
    "/api/v1/client/servers/{server}/snapshots/{snapshot}/restore": {
      "post": {
        "summary": "Restaurer le serveur depuis un instantané. Répond 202 : le serveur passe en restauration (démarrage, fichiers, SFTP refusés), s'arrête, puis l'agent prend un instantané de sûreté et recopie le dossier ; le serveur reste arrêté. Abandonné et rendu s'il ne s'arrête pas en 3 min. Refusé (409) sur un serveur occupé, quand l'agent se tait ou sous le seuil d'espace libre.",
        "description": "Clé personnelle. Portée requise : `snapshots.restore`.",
        "tags": [
          "Instantanés"
        ],
        "operationId": "client-post-servers-by-server-snapshots-by-snapshot-restore",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "required": true,
            "description": "Valeur de « snapshot ».",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": [
              "snapshots.restore"
            ]
          }
        ],
        "x-gd-scope": "snapshots.restore",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « snapshots.restore », ou refusé pour la ressource visée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-gd_session",
        "description": "Cookie de session du panel. Posé à la connexion, illisible par un script, et donc inutilisable depuis un client tiers."
      },
      "clientKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé personnelle, créée depuis « Compte → Clés d'API ». Elle agit au nom de son propriétaire et ne porte jamais plus que les portées qu'on lui a données."
      },
      "applicationKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé de plateforme, créée depuis l'administration. Elle n'agit au nom de personne : elle provisionne, suspend et résilie pour le compte d'un système tiers."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "Refus ou erreur, au format Problem Details (RFC 9457).",
        "properties": {
          "title": {
            "type": "string",
            "description": "Ce qui s'est passé, en une phrase."
          },
          "status": {
            "type": "integer",
            "description": "Le code HTTP, répété dans le corps."
          },
          "detail": {
            "type": "string",
            "description": "Le détail utile, quand il y en a un."
          }
        },
        "required": [
          "title",
          "status"
        ]
      }
    }
  }
}